+++ cli/src/operation_templater.rs
@@ -12,6 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+//! Templates for `jj op log`.
+//!
+//! See <https://jj-vcs.github.io/jj/latest/templates/#operation-keywords>.
+
 use std::any::Any;
 use std::cmp::Ordering;
 use std::collections::HashMap;
@@ -51,6 +55,7 @@ pub trait OperationTemplateLanguageExtension {
     fn build_cache_extensions(&self, extensions: &mut ExtensionsMap);
 }
 
+/// Templates for `jj op log`.
 pub struct OperationTemplateLanguage {
     repo_loader: RepoLoader,
     current_op_id: Option<OperationId>,