+++ lib/src/default_index/revset_engine.rs
@@ -1195,7 +1195,7 @@ fn build_predicate_fn(
             box_pure_predicate_fn(move |index, pos| {
                 let entry = index.entry_by_pos(pos);
                 let commit = store.get_commit(&entry.commit_id())?;
-                Ok(pattern.matches(commit.description()))
+                Ok(pattern.is_match(commit.description()))
             })
         }
         RevsetFilterPredicate::Subject(pattern) => {
@@ -1203,7 +1203,7 @@ fn build_predicate_fn(
             box_pure_predicate_fn(move |index, pos| {
                 let entry = index.entry_by_pos(pos);
                 let commit = store.get_commit(&entry.commit_id())?;
-                Ok(pattern.matches(commit.description().lines().next().unwrap_or_default()))
+                Ok(pattern.is_match(commit.description().lines().next().unwrap_or_default()))
             })
         }
         RevsetFilterPredicate::AuthorName(pattern) => {
@@ -1211,7 +1211,7 @@ fn build_predicate_fn(
             box_pure_predicate_fn(move |index, pos| {
                 let entry = index.entry_by_pos(pos);
                 let commit = store.get_commit(&entry.commit_id())?;
-                Ok(pattern.matches(&commit.author().name))
+                Ok(pattern.is_match(&commit.author().name))
             })
         }
         RevsetFilterPredicate::AuthorEmail(pattern) => {
@@ -1219,7 +1219,7 @@ fn build_predicate_fn(
             box_pure_predicate_fn(move |index, pos| {
                 let entry = index.entry_by_pos(pos);
                 let commit = store.get_commit(&entry.commit_id())?;
-                Ok(pattern.matches(&commit.author().email))
+                Ok(pattern.is_match(&commit.author().email))
             })
         }
         RevsetFilterPredicate::AuthorDate(expression) => {
@@ -1236,7 +1236,7 @@ fn build_predicate_fn(
             box_pure_predicate_fn(move |index, pos| {
                 let entry = index.entry_by_pos(pos);
                 let commit = store.get_commit(&entry.commit_id())?;
-                Ok(pattern.matches(&commit.committer().name))
+                Ok(pattern.is_match(&commit.committer().name))
             })
         }
         RevsetFilterPredicate::CommitterEmail(pattern) => {
@@ -1244,7 +1244,7 @@ fn build_predicate_fn(
             box_pure_predicate_fn(move |index, pos| {
                 let entry = index.entry_by_pos(pos);
                 let commit = store.get_commit(&entry.commit_id())?;
-                Ok(pattern.matches(&commit.committer().email))
+                Ok(pattern.is_match(&commit.committer().email))
             })
         }
         RevsetFilterPredicate::CommitterDate(expression) => {
@@ -1403,7 +1403,7 @@ fn match_lines<'a, 'b>(
     text.split_inclusive(|b| *b == b'\n').filter(|line| {
         let line = line.strip_suffix(b"\n").unwrap_or(line);
         // TODO: add .matches_bytes() or .to_bytes_matcher()
-        str::from_utf8(line).is_ok_and(|line| pattern.matches(line))
+        str::from_utf8(line).is_ok_and(|line| pattern.is_match(line))
     })
 }